Fixes some implicit type cast warnings

+++ fluid/Fl_Window_Type.cxx
@@ -109,7 +109,7 @@ void grid_cb(Fl_Int_Input *i, long v) {
 void default_widget_size_cb(Fl_Round_Button *b, long size) {
   // Update the "normal" text size of new widgets...
   b->setonly();
-  Fl_Widget_Type::default_size = size;
+  Fl_Widget_Type::default_size = (int)size;
   fluid_prefs.set("widget_size", Fl_Widget_Type::default_size);
 }
